refactor: redo huffman__generate_code
Several changes have been done to fix a memory error that appeared when reallocating memory each time a new node was added (since realloc does not ensure that the memory of the previous items is the same, and pointers were used to identify the lchild and rchild). The changes have included: - Instead of pointers, store the index of the lchild and rchild in the array. - Also, in huffman_codes, store directly the string (and its n of significant bits) for it to be easier to encode the message later